Models for links
Also, split the Post#add_term functionality into a Termable module, so
that we can mix it into other models (like Link) that can also be
attached to terms.
